GHDL: install library products into subdirs
This is consistent with other simulators and fixes
the

  clean_libraries

routine which removes library subdirs but not the
ghdl-created files in the CWD.

Also: create subdir when building the library not
after deleting it. This would have ghdl fail the
first time around and may confuse scripts that
use a presence test of a library subdir to check
whether there is a need to rebuild the library.